CMake Project Setup
Setting up the basic structure of the cmake project I am going to use
for implementing the renderer:
In the root of src is just the main.cpp located, implementing the
entrypoint into the application, but not more. The submodules Renderer
and Utils are containing the actual rendering code and obviously some
utilities required along the way.

Right now all files are just containing some dummy code to test the
setup.